[XLA:TPU] Use caching in CanShareOperandBufferWithUser.
Currently buffer assignment repeatedly calls CanShareOperandBufferWithUser. Each time this calls `GetInPlaceInputOutputPairs(user)`, then iterates through the pairs and checks the ones with the provided user_index. To avoid this repeated work, the first time we call `GetInPlaceInputOutputPairs(user)` we split and cache the values by user_index, avoiding duplicate work in subsequent calls.

We also cache the uses of the operand and speedup the HLO use check: converting the the inner for-loop into a lookup in a hash set.

PiperOrigin-RevId: 807400610